requirements.txt
boto3==1.40.7
boto3-stubs==1.40.7
botocore==1.40.7
glfw==2.9.0
jmespath==1.0.1
networkx==3.5
numpy==2.3.2
pygame==2.6.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
s3transfer==0.13.1
six==1.17.0
urllib3==2.5.0
